+++ components/updater/api/src/main/java/com/flipperdevices/updater/model/UpdateErrorType.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.updater.model
+
+enum class UpdateErrorType {
+    NO_INTERNET,
+    UNABLE_TO_SERVER,
+    NO_SD_CARD
+}