+++ CHANGELOG.md
@@ -7,9 +7,10 @@
 - [Feature] New error placeholder in FW Update
 - [Feature] Redesign Menu for choose firmware channel
 - [Feature] Add Low Battery dialog
-- [Feature] Test continue with deeplink in google play
+- [Feature] Show when flipper charging
+- [BUGFIX] Fix incorrect corner radius for battery icon
 
-# 1.1.2B
+# 1.1.2
 
 - [BUGFIX] Fix crash with custom firmware version
 - [BUGFIX] Fix crash when key file not exist